fix: get rid lockup in task locals
